+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 * LogRecord.toString() changed to use platform-specific line separators.
 * LogRecord.equals(Object): correctly handle null and other types.
+* RecordList.equals(Object): correctly handle null and other types.
 
 ### CI/CD
 * Integrated SpotBugs into build process.